Add language-specific pages and component
