+++ app/pl/page.tsx
@@ -1,0 +1,7 @@
+"use client"
+
+import TestimonialForm from "@/components/TestimonialForm"
+
+export default function Page() {
+  return <TestimonialForm language="pl" />
+}
